Moved graph storage and scrolled frame class into own files.
